pyproject.toml
[tool.poetry]
name = "Gate Keeper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-cors = "^5.0.0"
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
boto3 = "^1.35.56"
pyjwt = "^2.9.0"
cryptography = "^43.0.3"
flask-wtf = "^1.2.2"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]
isort = "^5.13.2"
black = "^24.10.0"
flake8 = "^7.1.1"
pytest-flask = "^1.3.0"
pytest-cov = "^5.0.0"
pytest = "^8.3.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
